rename and move components in web
