+++ apps/web/src/app/auth/(master-password)/login-master/page.tsx
@@ -1,4 +1,4 @@
-import LoginMasterForm from "@keyguard/web/components/auth/MasterPassword/LoginMasterForm";
+import LoginMasterForm from "@keyguard/web/app/auth/components/login-master-form";
 
 import AuthLayout from "../../auth.layout";
