+++ apps/web/src/app/auth/components/auth-token-handler.tsx
