+++ apps/web/src/components/sidebar/menu-items.ts
